+++ yash-env/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SystemEx::set_blocking`
 - `job::ProcessResult`
 - `job::ProcessResult::{exited, is_stopped}`
+- `impl From<job::ProcessResult> for ExitStatus`
 - `job::ProcessState::Halted`
 - `job::ProcessState::{stopped, exited, is_stopped}`
 - `impl From<trap::Condition> for stack::Frame`